+++ warp/tests/test_examples.py
@@ -20,9 +20,6 @@ override example defaults so the example can run in less than ten seconds.
 Use {"usd_required": True} and {"torch_required": True} to skip running the test
 if usd-core or torch are not found in the Python environment.
 
-Use "cutlass_required": True} to skip the test if Warp needs to be built with
-CUTLASS.
-
 Use the "num_frames" and "train_iters" keys to control the number of steps.
 
 Use "test_timeout" to override the default test timeout threshold of 300 seconds.
@@ -44,7 +41,7 @@ from warp.tests.unittest_utils import (
 )
 from warp.utils import check_p2p
 
-wp.init()  # For wp.context.runtime.core.is_cutlass_enabled()
+wp.init()  # For wp.context.runtime.core.is_debug_enabled()
 
 
 def _build_command_line_options(test_options: Dict[str, Any]) -> list:
@@ -110,10 +107,6 @@ def add_example_test(
         usd_required = options.pop("usd_required", False)
         if usd_required and not USD_AVAILABLE:
             test.skipTest("Requires usd-core")
-
-        cutlass_required = options.pop("cutlass_required", False)
-        if cutlass_required and not wp.context.runtime.core.is_cutlass_enabled():
-            test.skipTest("Warp was not built with CUTLASS support")
 
         # Find the current Warp cache
         warp_cache_path = wp.config.kernel_cache_dir
@@ -290,19 +283,6 @@ add_example_test(
     name="optim.example_trajectory",
     devices=test_devices,
     test_options={"headless": True, "train_iters": 50},
-)
-# NOTE: This example uses CUTLASS and will run orders of magnitude slower when Warp is built in debug mode
-add_example_test(
-    TestOptimExamples,
-    name="optim.example_walker",
-    devices=test_devices,
-    test_options={"usd_required": True},
-    test_options_cuda={
-        "train_iters": 1 if warp.context.runtime.core.is_debug_enabled() else 3,
-        "num_frames": 1 if warp.context.runtime.core.is_debug_enabled() else 60,
-        "cutlass_required": True,
-    },
-    test_options_cpu={"train_iters": 1, "num_frames": 30},
 )
 add_example_test(
     TestOptimExamples,